Preserve explicit FP32 LoRA parameters through model precision wrapping

+++ src/megatron/bridge/peft/lora_layers.py
@@ -465,7 +465,7 @@ class LinearAdapter(nn.Module):
         # pylint: disable=C0115,C0116
         if self.dropout_position == "pre":
             x = self.dropout(x)
-        lora_res = self.linear_out(self.linear_in(x))
+        lora_res = self.linear_out(self.linear_in(x.to(self.linear_in.weight.dtype)))
         lora_res = lora_res * self.scale
         if self.dropout_position == "post":
             lora_res = self.dropout(lora_res)
